ext-service/nlp-processing/requirements.txt
toml
requests
tornado
pynsq